third_party/blink/web_tests/platform/mac-mac10.10/http/tests/devtools/indexeddb/resources-panel-expected.txt

CONSOLE MESSAGE: line 49: InspectorTest.IndexedDB_callback1
CONSOLE MESSAGE: line 49: InspectorTest.IndexedDB_callback2
CONSOLE MESSAGE: line 49: InspectorTest.IndexedDB_callback3
Tests IndexedDB tree element on resources panel.

Expanded IndexedDB tree element.
Dumping IndexedDB tree:
    (empty)
Creating database.
Created database.
Refreshing.
Refreshed.
Dumping IndexedDB tree:
    database: testDatabase - http://127.0.0.1:8000
        Object store: testObjectStore
            Index: testIndexName
Navigating to another security origin.
Navigated to another security origin.
Dumping IndexedDB tree:
    (empty)
Navigating back.
Navigated back.
Refreshing.
Refreshed.
Dumping IndexedDB tree:
    database: testDatabase - http://127.0.0.1:8000
        Object store: testObjectStore
            Index: testIndexName
Deleting database.
Deleted database.
Refreshing.
Refreshed.
Dumping IndexedDB tree:
    (empty)
